refactor safe extract method to fix issue 6215
